Utilize the prezto framework for customizations
Symlink the contrib directory to one being tracked by the dotfiles repo. The contrib is ignored by prezto so this is shouldn’t cause any merge conflicts when pulling the upstream.
I’m loading a user module which will serve as my site-lisp of sorts, but with the added benefit of version control which can remain privately managed .